fixUserService,fixRestaurantService.remakeRestaurantAddComment,remakeRestaurantGetComment

+++ app/Services/UserFavorite.php
@@ -28,7 +28,12 @@ class UserFavorite
     public function get($userId)
     {
         try {
+            if (!isset($userId)) {
+                throw new Exception('資料缺失');
+            }
             return User_favorite::where('uid', '=', $userId)->get()->toArray();
+        } catch (Exception $e) {
+            throw new Exception($e->getMessage());
         } catch (Throwable $e) {
             throw new Exception("user_favorite_service_err:" . 500);
         }
@@ -36,7 +41,12 @@ class UserFavorite
     public function delete($userId, $rid)
     {
         try {
+            if (!isset($userId) || !isset($rid)) {
+                throw new Exception('資料缺失');
+            }
             return User_favorite::where('uid', '=', $userId)->where('rid', '=', $rid)->delete();
+        } catch (Exception $e) {
+            throw new Exception($e->getMessage());
         } catch (Throwable $e) {
             throw new Exception("user_favorite_service_err:" . 500);
         }
